CRYP-4449: Move solana/web3 from keyring.ts to solana.ts
